requirements.txt
aiohttp
httpx
pytest-asyncio
requests
websockets
